refactor: sidebar components and introduce custom sidebar hook for improved state management
