+++ src/components/dashboard/Sidebar.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import { useState } from 'react';
 import logoIcon from '@/images/logo.svg';
 import {
   Sidebar as UISidebar,
@@ -8,39 +7,23 @@ import {
   SidebarGroup,
   SidebarGroupContent,
   SidebarHeader,
-  useSidebar,
 } from '@/components/ui/sidebar';
 import { Separator } from '@/components/ui/separator';
 import Image from 'next/image';
-import SidebarMenuGroup from './SidebarMenuGroup';
-import { items } from './SidebarItems';
-import { useIsMobile } from '@/hooks/use-mobile';
+import { SidebarMenuGroup } from '@/components/dashboard/SidebarMenuGroup';
+import { items } from '@/data/SidebarItems';
+import { useCustomSidebar } from '@/hooks/useCustomSidebar';
 
 const group1 = items.slice(0, 6);
 const group2 = items.slice(6);
 
-const Sidebar = () => {
-  const [activeTitle, setActiveTitle] = useState(items[0].title);
-  const { toggleSidebar, openMobile } = useSidebar();
-  const isMobile = useIsMobile();
-
-  const handleSidebarToggle = () => {
-    if (isMobile) {
-      toggleSidebar();
-    }
-  };
+export const Sidebar = () => {
+  const { activeTitle, setActiveTitle, handleSidebarToggle, openMobile, isMobile } = useCustomSidebar();
 
   return (
     <UISidebar>
       <SidebarHeader className="flex flex-row gap-3 pt-[37px] pl-5 md:pl-8 pb-0 items-center mb-[59px]">
-        <Image
-          src={logoIcon}
-          alt="Logo"
-          width={32}
-          height={32}
-          onClick={handleSidebarToggle}
-          className="cursor-pointer md:cursor-auto"
-        />
+        <Image src={logoIcon} alt="Logo" width={32} height={32} onClick={handleSidebarToggle} className="cursor-pointer md:cursor-auto" />
         {((isMobile && openMobile) || !isMobile) && (
           <h1 className="text-[19px] tracking-[0.4px] font-bold">
             Dashboard Kit
@@ -67,5 +50,3 @@ const Sidebar = () => {
     </UISidebar>
   );
 };
-
-export default Sidebar;
